fix: Move the CurrentStatus enumeration to the FlightAssignments package
