+++ src/main/java/acme/entities/flightAssignments/CurrentStatus.java
@@ -1,5 +1,5 @@
 
-package acme.entities.flights;
+package acme.entities.flightAssignments;
 
 public enum CurrentStatus {
 	CONFIRMED, PENDING, CANCELLED
