+++ src/main/java/acme/entities/flightAssignments/FlightAssignment.java
@@ -17,7 +17,6 @@ import acme.client.components.validation.Mandatory;
 import acme.client.components.validation.Optional;
 import acme.client.components.validation.ValidMoment;
 import acme.client.components.validation.ValidString;
-import acme.entities.flights.CurrentStatus;
 import acme.entities.legs.Leg;
 import acme.realms.FlightCrewMember;
 import lombok.Getter;
